requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
bleak==0.22.3
bluepy==1.3.0
certifi==2025.4.26
charset-normalizer==3.4.1
dbus-fast==2.44.1
frozenlist==1.6.0
future==1.0.0
idna==3.10
influxdb-client==1.48.0
lywsd02==0.0.9
mac-vendor-lookup==0.1.12
multidict==6.4.3
propcache==0.3.1
pycryptodome==3.22.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
reactivex==4.0.4
requests==2.32.3
setuptools==79.0.1
six==1.17.0
tabulate==0.9.0
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
